pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "media-lavarage-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "anthropic>=0.25",
    "openai>=1.30",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["watchfiles>=0.21"]

[project.scripts]
media-agent = "main:main"

[tool.setuptools.packages.find]
where = ["."]
